+++ src/SatoshiRig/__init__.py
@@ -1,3 +1,7 @@
+
+import logging
+logging.basicConfig(level=logging.INFO, filename='miner.log', filemode='w', format='%(asctime)s - %(levelname)s - %(message)s')
+
 __all__ = [
     "cli",
     "miner",